Major update

+++ resources/views/participants/index.blade.php
@@ -17,7 +17,7 @@
 
                     @if (count($competitionA) <= 4)
                         <script> minError = true;</script>
-                    @else
+                    @elseif (count($competitionA) > 8)
                         <script> maxError = true;</script>
                     @endif
                     <table class="table table-striped">
@@ -39,12 +39,18 @@
                                     <td>{{ $participant->school }}</td>
                                     <td>{{ $participant->class }}</td>
                                     <td>{{ $participant->level }}</td>
-                                    <td><a class="btn btn-danger btn-xs" name="delete" href="/participants/delete/{{ $participant->id }}">Verwijder</a></td>
+                                    @if (!$startedA)
+                                        <td><a class="btn btn-danger btn-xs" name="delete" href="/participants/delete/{{ $participant->id }}">Verwijder</a></td>
+                                    @endif
                                  </tr>
                             @endforeach
                         </tbody>
                     </table>
-                    <a href="/scheme/store" name="createCompetition" data-competition="a" class="btn btn-success">Wedstrijdschema poule A aanmaken</a>
+                    @if (!$startedA)
+                        <a href="/scheme/generate" name="createCompetition" data-competition="0" class="btn btn-success">Wedstrijdschema poule A aanmaken</a>
+                    @else
+                        <a href="" name="deleteCompetition" data-competition="0" class="btn btn-danger">Wedstrijdschema poule A resetten</a>
+                    @endif
                 @else
                     <h3>Er zijn geen aanmeldingen voor deze poule</h3>
                 @endif
@@ -54,6 +60,13 @@
             </div>
             <div class="tab-pane fade" id="b" role="tabpanel" aria-labelledby="b-tab">
                 @if (!$competitionB->isEmpty())
+                    <script> var minError = false; var maxError = false;</script>
+
+                    @if (count($competitionB) <= 4)
+                        <script> minError = true;</script>
+                    @elseif (count($competitionB) > 8)
+                        <script> maxError = true;</script>
+                    @endif
                     <table class="table table-striped">
                         <thead>
                         <tr>
@@ -66,19 +79,25 @@
                         </tr>
                         </thead>
                         <tbody>
-                            @foreach($competitionB as $key => $participant)
-                                <tr>
-                                    <td>{{$key + 1}}</td>
-                                    <td>{{$participant->name}}</td>
-                                    <td>{{ $participant->school }}</td>
-                                    <td>{{ $participant->class }}</td>
-                                    <td>{{ $participant->level }}</td>
+                        @foreach($competitionB as $key => $participant)
+                            <tr>
+                                <td>{{$key + 1}}</td>
+                                <td>{{$participant->name}}</td>
+                                <td>{{ $participant->school }}</td>
+                                <td>{{ $participant->class }}</td>
+                                <td>{{ $participant->level }}</td>
+                                @if (!$startedB)
                                     <td><a class="btn btn-danger btn-xs" name="delete" href="/participants/delete/{{ $participant->id }}">Verwijder</a></td>
-                                </tr>
-                            @endforeach
+                                @endif
+                            </tr>
+                        @endforeach
                         </tbody>
                     </table>
-                    <a href="" name="createCompetition" data-competition="b" class="btn btn-success">Wedstrijdschema poule B aanmaken</a>
+                    @if (!$startedB)
+                        <a href="/scheme/generate" name="createCompetition" data-competition="1" class="btn btn-success">Wedstrijdschema poule B aanmaken</a>
+                    @else
+                        <a href="" name="deleteCompetition" data-competition="1" class="btn btn-danger">Wedstrijdschema poule B resetten</a>
+                    @endif
                 @else
                     <h3>Er zijn geen aanmeldingen voor deze poule</h3>
                 @endif
